Remove csv reader
