+++ dataset/copy-node-property-test/vPerson.csv
@@ -1,12 +1,12 @@
 0,
 1,"ozwhvnetnq"
-2,"kuk,qg\\nrspmk"
-3,"wmz,1234\lamo"
+2,"kuk,qgnrspmk"
+3,"wmz,1234lamo"
 4,"tudoojdduf"
 5,"qifidjufri"
 6,"gqpnpbdmrb"
 7,"dgzbiqjkaz"
-8,"ebf,,uq\buqma"
+8,"ebf,,uqbuqma"
 9,"rwhnybogfy"
 10,"enqpnymvdb"
 11,"axgwwhhohf"
