+++ src/common/copier_config/CMakeLists.txt
@@ -1,0 +1,7 @@
+add_library(kuzu_common_copier_config
+        OBJECT
+        copier_config.cpp)
+
+set(ALL_OBJECT_FILES
+        ${ALL_OBJECT_FILES} $<TARGET_OBJECTS:kuzu_common_copier_config>
+        PARENT_SCOPE)
