+++ src/include/binder/copy/bound_copy.h
@@ -6,7 +6,7 @@
 
 #include "binder/bound_statement.h"
 #include "catalog/catalog_structs.h"
-#include "common/csv_reader/csv_reader.h"
+#include "common/copier_config/copier_config.h"
 
 namespace kuzu {
 namespace binder {
